Relicense to GPLv3
With the release of Slim Theme Manager, we have received some feedback.
Even though this feedback was in a crude form, we have decided to listen to it.

The first step of this is relicensing all our components to the GPLv3 license. This
is done in effort to re-open ties with other theme providers.
We intend to contribute our findings to them, and hope that they will contribute
their findings back to us, so this community can once again be unified.

Due to the componented nature of our system, this will not cause many issues
for those wanting to write a backend under different licenses.

Also, some minor formatting edits.

Change-Id: I02e454f038e50826dfd16080a582260310e59531